feat: tooltip
